repositories: refactor mappings
The project was being used just for the mappings, while actual
repositories were at BusinessLogic.
