+++ site/DFM.BusinessLogic/DFM.BusinessLogic.csproj
@@ -84,6 +84,21 @@
     <Compile Include="Repositories\CategoryRepository.cs" />
     <Compile Include="Repositories\ConfigRepository.cs" />
     <Compile Include="Repositories\ContractRepository.cs" />
+    <Compile Include="Repositories\Mappings\AcceptanceMap.cs" />
+    <Compile Include="Repositories\Mappings\AccountMap.cs" />
+    <Compile Include="Repositories\Mappings\CategoryMap.cs" />
+    <Compile Include="Repositories\Mappings\ConfigMap.cs" />
+    <Compile Include="Repositories\Mappings\ContractMap.cs" />
+    <Compile Include="Repositories\Mappings\DetailMap.cs" />
+    <Compile Include="Repositories\DFMRepositoryException.cs" />
+    <Compile Include="Repositories\Mappings\MonthMap.cs" />
+    <Compile Include="Repositories\Mappings\MoveMap.cs" />
+    <Compile Include="Repositories\Mappings\ScheduleMap.cs" />
+    <Compile Include="Repositories\Mappings\SecurityMap.cs" />
+    <Compile Include="Repositories\Mappings\SummaryMap.cs" />
+    <Compile Include="Repositories\Mappings\TicketMap.cs" />
+    <Compile Include="Repositories\Mappings\UserMap.cs" />
+    <Compile Include="Repositories\Mappings\YearMap.cs" />
     <Compile Include="ServiceAccess.cs" />
     <Compile Include="Repositories\DetailRepository.cs" />
     <Compile Include="Repositories\MonthRepository.cs" />
