+++ site/DFM.BusinessLogic/Repositories/Mappings/SummaryMap.cs
@@ -2,7 +2,7 @@
 using FluentNHibernate.Automapping;
 using FluentNHibernate.Automapping.Alterations;
 
-namespace DFM.Repositories.Mappings
+namespace DFM.BusinessLogic.Repositories.Mappings
 {
 	public class SummaryMap : IAutoMappingOverride<Summary>
 	{
