+++ site/DFM.MVC/DFM.MVC.csproj
@@ -496,10 +496,6 @@
       <Project>{A747268B-CE55-4577-8D72-5405A2E5866C}</Project>
       <Name>DFM.Multilanguage</Name>
     </ProjectReference>
-    <ProjectReference Include="..\DFM.Repositories\DFM.Repositories.csproj">
-      <Project>{40054C79-2502-4A01-AF87-8AB2489F8A0D}</Project>
-      <Name>DFM.Repositories</Name>
-    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Assets\Resources\General.resx">
