+++ site/DFM.MVC/Global.asax.cs
@@ -10,13 +10,13 @@ using System.Web.Routing;
 using Keon.Util.Exceptions;
 using DFM.Authentication;
 using DFM.BusinessLogic;
+using DFM.BusinessLogic.Repositories.Mappings;
 using DFM.Email;
 using DFM.Entities;
 using DFM.Generic;
 using DFM.MVC.Controllers;
 using DFM.MVC.Helpers;
 using DFM.MVC.Helpers.Global;
-using DFM.Repositories.Mappings;
 using Keon.MVC.Cookies;
 using Keon.NHibernate.Base;
 using log4net.Config;
