--- site/DFM.Repositories/DFM.Repositories.csproj
@@ -1,119 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProductVersion>8.0.30703</ProductVersion>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{40054C79-2502-4A01-AF87-8AB2489F8A0D}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>DFM.Repositories</RootNamespace>
-    <AssemblyName>DFM.Repositories</AssemblyName>
-    <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <TargetFrameworkProfile />
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <Prefer32Bit>false</Prefer32Bit>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <Prefer32Bit>false</Prefer32Bit>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="..\DFM.Generic\Properties\GeneralAssemblyInfo.cs">
-      <Link>Properties\GeneralAssemblyInfo.cs</Link>
-    </Compile>
-    <Compile Include="DFMRepositoryException.cs" />
-    <Compile Include="Mappings\AccountMap.cs" />
-    <Compile Include="Mappings\CategoryMap.cs" />
-    <Compile Include="Mappings\ConfigMap.cs" />
-    <Compile Include="Mappings\AcceptanceMap.cs" />
-    <Compile Include="Mappings\ContractMap.cs" />
-    <Compile Include="Mappings\TicketMap.cs" />
-    <Compile Include="Mappings\DetailMap.cs" />
-    <Compile Include="Mappings\MonthMap.cs" />
-    <Compile Include="Mappings\MoveMap.cs" />
-    <Compile Include="Mappings\ScheduleMap.cs" />
-    <Compile Include="Mappings\SecurityMap.cs" />
-    <Compile Include="Mappings\SummaryMap.cs" />
-    <Compile Include="Mappings\UserMap.cs" />
-    <Compile Include="Mappings\YearMap.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\DFM.Authentication\DFM.Authentication.csproj">
-      <Project>{84AD52CD-61AA-4F63-B26F-92392493E4DC}</Project>
-      <Name>DFM.Authentication</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\DFM.BusinessLogic\DFM.BusinessLogic.csproj">
-      <Project>{E749C813-2963-49E6-9D0A-7E6F165EF929}</Project>
-      <Name>DFM.BusinessLogic</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\DFM.Entities\DFM.Entities.csproj">
-      <Project>{288B6282-630A-4D92-BF07-717774BFB2BD}</Project>
-      <Name>DFM.Entities</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\DFM.Generic\DFM.Generic.csproj">
-      <Project>{FD6843E9-6DC1-464A-A6D6-74AE215BD16D}</Project>
-      <Name>DFM.Generic</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <Reference Include="Antlr3.Runtime, Version=3.5.0.2, Culture=neutral, PublicKeyToken=eb42632606e9261f, processorArchitecture=MSIL">
-      <HintPath>..\packages\Antlr3.Runtime.3.5.1\lib\net40-client\Antlr3.Runtime.dll</HintPath>
-    </Reference>
-    <Reference Include="FluentNHibernate, Version=2.1.2.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\FluentNHibernate.2.1.2\lib\net461\FluentNHibernate.dll</HintPath>
-    </Reference>
-    <Reference Include="Iesi.Collections, Version=4.0.0.4000, Culture=neutral, PublicKeyToken=aa95f207798dfdb4, processorArchitecture=MSIL">
-      <HintPath>..\packages\Iesi.Collections.4.0.4\lib\net461\Iesi.Collections.dll</HintPath>
-    </Reference>
-    <Reference Include="Keon.NHibernate, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Keon.NHibernate.1.0.0\lib\net47\Keon.NHibernate.dll</HintPath>
-    </Reference>
-    <Reference Include="Keon.Util, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Keon.Util.1.0.0\lib\net47\Keon.Util.dll</HintPath>
-    </Reference>
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="NHibernate, Version=5.2.0.0, Culture=neutral, PublicKeyToken=aa95f207798dfdb4, processorArchitecture=MSIL">
-      <HintPath>..\packages\NHibernate.5.2.5\lib\net461\NHibernate.dll</HintPath>
-    </Reference>
-    <Reference Include="Remotion.Linq, Version=2.2.0.0, Culture=neutral, PublicKeyToken=fee00910d6e5f53b, processorArchitecture=MSIL">
-      <HintPath>..\packages\Remotion.Linq.2.2.0\lib\net45\Remotion.Linq.dll</HintPath>
-    </Reference>
-    <Reference Include="Remotion.Linq.EagerFetching, Version=2.2.0.0, Culture=neutral, PublicKeyToken=fee00910d6e5f53b, processorArchitecture=MSIL">
-      <HintPath>..\packages\Remotion.Linq.EagerFetching.2.2.0\lib\net45\Remotion.Linq.EagerFetching.dll</HintPath>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.Configuration" />
-    <Reference Include="System.ServiceModel" />
-    <Reference Include="System.Transactions" />
-    <Reference Include="System.Web" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="app.config" />
-    <None Include="packages.config" />
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
-</Project>
