--- site/DFM.Repositories/Properties/AssemblyInfo.cs
@@ -1,6 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("DFM.Repositories")]
-[assembly: AssemblyProduct("DFM.Repositories")]
-[assembly: Guid("ad44ea11-9138-4e9d-9d3d-81c2e7bc5873")]
