--- site/DFM.Repositories/packages.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Antlr3.Runtime" version="3.5.1" targetFramework="net47" />
-  <package id="FluentNHibernate" version="2.1.2" targetFramework="net471" />
-  <package id="Iesi.Collections" version="4.0.4" targetFramework="net471" />
-  <package id="Keon.NHibernate" version="1.0.0" targetFramework="net472" />
-  <package id="Keon.Util"  version="1.0.0" targetFramework="net472" />
-  <package id="NHibernate" version="5.2.5" targetFramework="net471" />
-  <package id="Remotion.Linq" version="2.2.0" targetFramework="net471" />
-  <package id="Remotion.Linq.EagerFetching" version="2.2.0" targetFramework="net471" />
-</packages>
