+++ site/DFM.Tests/BusinessLogic/BaseStep.cs
@@ -8,10 +8,10 @@ using DFM.Email;
 using DFM.Entities;
 using DFM.Entities.Enums;
 using DFM.Multilanguage;
-using DFM.Repositories.Mappings;
 using DFM.Tests.BusinessLogic.Helpers;
 using DFM.Tests.Helpers;
 using System.Text.RegularExpressions;
+using DFM.BusinessLogic.Repositories.Mappings;
 using Keon.Util.Extensions;
 using Keon.NHibernate.Base;
 using TechTalk.SpecFlow;
