+++ site/DFM.Tests/DFM.Tests.csproj
@@ -235,10 +235,6 @@
       <Project>{A747268B-CE55-4577-8D72-5405A2E5866C}</Project>
       <Name>DFM.Multilanguage</Name>
     </ProjectReference>
-    <ProjectReference Include="..\DFM.Repositories\DFM.Repositories.csproj">
-      <Project>{40054C79-2502-4A01-AF87-8AB2489F8A0D}</Project>
-      <Name>DFM.Repositories</Name>
-    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
